+++ src/Glob/Glob.cs
@@ -51,48 +51,11 @@ namespace GlobExpressions
                 var last = pathSegments.LastOrDefault();
                 var tail = (last == null) ? new string[0] : new[] { last };
 
-                if (IsMatch(_segments, 0, tail, 0))
+                if (GlobEvaluator.Eval(_segments, 0, tail, 0, _caseSensitive))
                     return true;
             }
 
-            return IsMatch(_segments, 0, pathSegments, 0);
-        }
-
-        private bool IsMatch(Segment[] pattern, int patternIndex, string[] input, int inputIndex)
-        {
-            while (true)
-            {
-                if (inputIndex == input.Length)
-                    return patternIndex == pattern.Length;
-
-                // we have a input to match but no pattern to match against so we are done.
-                if (patternIndex == pattern.Length)
-                    return false;
-
-                var inputHead = input[inputIndex];
-                var patternHead = pattern[patternIndex];
-
-                switch (patternHead)
-                {
-                    case DirectoryWildcard _:
-                        // return all consuming the wildcard
-                        return IsMatch(pattern, patternIndex + 1, input, inputIndex) // 0 matches
-                               || IsMatch(pattern, patternIndex + 1, input, inputIndex + 1) // 1 match
-                               || IsMatch(pattern, patternIndex, input, inputIndex + 1); // more
-
-                    case Root root when inputHead == root.Text:
-                        patternIndex++;
-                        inputIndex++;
-                        continue;
-
-                    case DirectorySegment dir when dir.MatchesSegment(inputHead, _caseSensitive):
-                        patternIndex++;
-                        inputIndex++;
-                        continue;
-                }
-
-                return false;
-            }
+            return GlobEvaluator.Eval(_segments, 0, pathSegments, 0, _caseSensitive);
         }
     }
 }