+++ src/Glob/Token.cs
@@ -5,16 +5,22 @@ using System.Text;
 
 namespace GlobExpressions
 {
-    class Token
+    internal struct Token
     {
-        public Token(TokenKind kind, string spelling)
+        private readonly string _source;
+        private readonly int _startIndex;
+        private readonly int _length;
+
+        public Token(TokenKind kind, string source, int startIndex, int length)
         {
+            _source = source;
+            _startIndex = startIndex;
+            _length = length;
             this.Kind = kind;
-            this.Spelling = spelling;
         }
 
         public TokenKind Kind { get; }
-        public string Spelling { get; }
+        public string Spelling => _source.Substring(_startIndex, _length);
 
         public override string ToString()
         {