+++ test/Glob.Tests/TestHelpers.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Text;
+
+namespace GlobExpressions.Tests
+{
+    public static class TestHelpers
+    {
+        public static readonly string SourceRoot = Environment.GetEnvironmentVariable("APPVEYOR_BUILD_FOLDER") ?? Path.Combine("..", "..", "..", "..", "..");
+    }
+}